requirements.txt
yt-dlp>=2026.1.0
pyyaml>=6.0
# Optional (for programmatic AI calls, not required if using Claude Code skill):
# anthropic>=0.40.0
